requirements.txt
# 数据采集
akshare>=1.12.0
baostock>=0.8.8

# Web框架
fastapi>=0.109.0
uvicorn>=0.27.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 认证
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# 邮件
aiosmtplib>=3.0.0

# 工具
python-dateutil>=2.8.0
pydantic>=2.5.0
httpx>=0.26.0

# 报告生成
Pillow>=10.0.0

# 拼音搜索
pypinyin>=0.49.0
